macvim: Fix compilation on Mac OS X 10.9
On Mavericks we need to explicitly include AvailabilityMacros.h to have
the version detection macros defined.

Fix 10.9 ruby framework detection and compilation

Mavericks ships with version 1.8 and 2.0 of the Ruby.framework, so we
must directly link the framework version matching the ruby-command.

This also means that ruby.h must no longer included via the framework
name, which has the nice side effect of allowing you to compile with
non-framework rubies, if you remove the formula's hardcoded RUBY_PATH.